+++ packages/client/src/play/index.js
@@ -1,0 +1,2 @@
+export { createLivePlayClient } from './live.js'
+export * from './schema.js'